chore: split up rambo
