+++ src/programs/DecoratorTest.ts
@@ -1,20 +1,20 @@
 import { Message } from "discord.js";
-import { Command, CommandHandler, DiscordEvent } from "../events/handler";
+import { Command, CommandHandler, DiscordEvent } from "../rambo";
 
 /**
  * Example of a stateless message handler (stateful: true missing in config).
  * Each time the command is run, a new instance is created so the counter resets.
  */
 @Command({
-  event: DiscordEvent.Message,
+  event: DiscordEvent.MESSAGE,
   description: "",
   trigger: "!test",
   channelNames: ["permanent-testing"],
 })
-export class DecoratorTest extends CommandHandler<DiscordEvent.Message> {
+export class DecoratorTest extends CommandHandler<DiscordEvent.MESSAGE> {
   called: number = 0;
 
-  handleEvent(message: Message): void {
+  handleEvent(): void {
     ++this.called;
     console.log(`Called handler 1 ${this.called} times`);
   }
@@ -25,13 +25,13 @@ export class DecoratorTest extends CommandHandler<DiscordEvent.Message> {
  * Since this is kept as a singleton, the increasing counter is maintained.
  */
 @Command({
-  event: DiscordEvent.Message,
+  event: DiscordEvent.MESSAGE,
   description: "",
   trigger: "!test",
   stateful: true,
   channelNames: ["bot-output"],
 })
-export class DecoratorTest2 extends CommandHandler<DiscordEvent.Message> {
+export class DecoratorTest2 extends CommandHandler<DiscordEvent.MESSAGE> {
   called: number = 0;
 
   handleEvent(message: Message): void {
