+++ src/rambo/README.md
@@ -1,0 +1,6 @@
+# Rambo
+
+Rambo is the core of YesBot, responsible for loading all commands and event handlers and calling the relevant ones when
+a new event comes in from Discord.
+
+TODO: Documentation of the decorator, supported event types, function signatures, etc
