+++ src/rambo/events/index.ts
@@ -1,0 +1,66 @@
+import {
+  addMessageHandler,
+  extractMessageInfo,
+  MessageEventHandlerOptions,
+  MessageHandlerFunction,
+} from "./message";
+import {
+  AddEventHandlerFunction,
+  DiscordEvent,
+  ExtractInfoFunction,
+} from "../types/base";
+import {
+  addReactionHandler,
+  extractReactionInfo,
+  ReactionEventHandlerOptions,
+  ReactionHandlerFunction,
+} from "./reactions";
+import { StringIndexedHIOCTree } from "../types/hioc";
+import { Message, MessageReaction, User } from "discord.js";
+
+export type EventHandlerOptions =
+  | MessageEventHandlerOptions
+  | ReactionEventHandlerOptions;
+
+export type HandlerFunction<T extends DiscordEvent> =
+  | MessageHandlerFunction<T>
+  | ReactionHandlerFunction<T>;
+
+export const addEventHandler: AddEventHandlerFunction<EventHandlerOptions> = (
+  options,
+  ioc,
+  tree
+) => {
+  switch (options.event) {
+    case DiscordEvent.MESSAGE:
+      return addMessageHandler(
+        options,
+        ioc,
+        tree as StringIndexedHIOCTree<DiscordEvent.MESSAGE>
+      );
+    case DiscordEvent.REACTION_ADD:
+    case DiscordEvent.REACTION_REMOVE:
+      return addReactionHandler(
+        options,
+        ioc,
+        tree as StringIndexedHIOCTree<
+          DiscordEvent.REACTION_ADD | DiscordEvent.REACTION_REMOVE
+        >
+      );
+  }
+};
+
+export const extractEventInfo: ExtractInfoFunction<DiscordEvent> = (
+  event,
+  ...args
+) => {
+  switch (event) {
+    case DiscordEvent.MESSAGE:
+      return extractMessageInfo(args[0] as Message);
+    case DiscordEvent.REACTION_ADD:
+    case DiscordEvent.REACTION_REMOVE:
+      return extractReactionInfo(args[0] as MessageReaction, args[1] as User);
+    default:
+      throw new Error("Could not extract info for event " + event);
+  }
+};
